requirements.txt
beautifulsoup4
pillow
requests
RPi.GPIO
schedule
spidev
urllib3
waveshare-epaper